refactor: move unreachable-lambda functions to isolatable files
- DrawerOneUIExtensions.kt (new): move setupHeaderAndNavRail and
  onNavigationSingleClick from DrawerUtils.kt; these functions contain
  invokeDynamic lambdas that Kover's annotatedBy cannot exclude because
  lambdas compile as static methods in the enclosing class, not as
  annotated methods. Moving to a dedicated file allows a classes()
  exclusion on *DrawerOneUIExtensionsKt* to cover the whole class.
- DeleteAppDataUtils.kt (new): same treatment for deleteAppDataAndExit
  (setOnClickListenerWithProgress + coroutine body), excluded via
  *DeleteAppDataUtilsKt*.
- CommonUtilsAboutActivity.kt: switch Play Core callbacks from lambdas
  to method references (::onUpdateAvailabilityFetched, ::onUpdateFetchFailed,
  ::onResumeUpdateCheck, ::onActivityResult) — method refs do not generate
  static lambda methods, eliminating those missed instructions.
- CommonUtilsAboutMeActivity.kt: extract handleShareApp, handleWriteEmail,
  onPlayStoreConfirmed as @VisibleForTesting internal functions and use
  method references for click/dialog listeners, making those paths
  directly testable and eliminating static lambda miss.
