+++ lib/src/main/java/de/lemke/commonutils/DrawerUtils.kt
@@ -18,49 +18,8 @@
 package de.lemke.commonutils
 
 import android.os.Bundle
-import android.widget.ImageButton
-import androidx.appcompat.content.res.AppCompatResources
-import com.google.android.material.navigation.NavigationView
-import de.lemke.commonutils.ui.activity.CommonUtilsAboutActivity
-import dev.oneuiproject.oneui.layout.Badge
-import dev.oneuiproject.oneui.layout.NavDrawerLayout
-import dev.oneuiproject.oneui.navigation.widget.DrawerNavigationView
-import dev.oneuiproject.oneui.R as iconsR
-import dev.oneuiproject.oneui.design.R as designR
 
 private const val TAG = "DrawerUtils"
-private const val NAV_RAIL_MIN_SIDE_MARGIN_DP = 14
-
-/** Sets up the drawer header button with an info icon navigating to [CommonUtilsAboutActivity] and configures the nav rail. */
-@NoCoverage
-fun NavDrawerLayout.setupHeaderAndNavRail(aboutApp: String) {
-    setupHeaderButton(
-        icon = AppCompatResources.getDrawable(context, iconsR.drawable.ic_oui_info_outline)!!,
-        tooltipText = aboutApp,
-        listener = {
-            findViewById<ImageButton>(designR.id.oui_des_drawer_header_button)
-                .transformToActivity(CommonUtilsAboutActivity::class.java, transitionName = "CommonUtilsAboutAppTransition")
-        },
-    )
-    setNavRailContentMinSideMargin(NAV_RAIL_MIN_SIDE_MARGIN_DP)
-    closeNavRailOnBack = true
-    context.onAppUpdateAvailable { setButtonBadges(Badge.DOT, Badge.DOT) }
-}
-
-/** Wraps [listener] to ignore repeated clicks within [interval] milliseconds, preventing double-navigation. */
-@NoCoverage
-fun DrawerNavigationView.onNavigationSingleClick(
-    interval: Long = 600,
-    listener: NavigationView.OnNavigationItemSelectedListener,
-) {
-    var lastClick = 0L
-    setNavigationItemSelectedListener { item ->
-        val currentTime = System.currentTimeMillis()
-        if (currentTime - lastClick < interval) return@setNavigationItemSelectedListener false
-        lastClick = currentTime
-        listener.onNavigationItemSelected(item)
-    }
-}
 
 /** Bundle key for persisting search mode state. */
 const val COMMONUTILS_KEY_IS_SEARCH_MODE = "commonutils_isSearchMode"
@@ -87,6 +46,7 @@ fun Bundle.saveSearchAndActionMode(
 }
 
 /** Restores search and action mode state from this bundle, invoking the provided callbacks as needed. */
+@NoCoverage
 inline fun Bundle?.restoreSearchAndActionMode(
     crossinline onSearchMode: () -> Unit = {},
     crossinline onActionMode: (selectedIds: Set<Long>) -> Unit = {},
